requirements.txt
alembic==1.5.7
atomicwrites==1.4.0
attrs==20.3.0
bleach==3.2.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorama==0.4.3
docutils==0.16
embedding-as-service-client==1.0.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
gunicorn==20.1.0
idna==2.8
IMDbPY==2021.4.18
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
keyring==21.4.0
lxml==4.6.3
Mako==1.1.4
MarkupSafe==1.1.1
numpy==1.20.1
oauthlib==3.1.0
packaging==20.9
pandas==1.2.4
pkginfo==1.5.0.1
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
Pygments==2.7.1
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.2
pytest-flask==1.2.0
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
pywin32-ctypes==0.2.0
pyzmq==22.0.3
readme-renderer==26.0
requests==2.21.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sklearn==0.0
SQLAlchemy==1.4.2
threadpoolctl==2.1.0
tmdbsimple==2.8.0
toml==0.10.2
tqdm==4.50.1
tweepy==3.10.0
twine==3.2.0
urllib3==1.24.3
webencodings==0.5.1
Werkzeug==1.0.1
wincertstore==0.2
zmq==0.0.0
